First pass at using Scalastyle
I just kind of went with rules that seemed right for the project. People
are free to suggest changes.

Currently Scalatest is failing to parse Free.scala which seems like a
bug in Scalastyle. The error message is "Free.scala: Expected token
RBRACKET but got Token(DOT,.,1985,.)". I believe the following code is
tripping it up:

```scala
case y: Gosub[S, x.C] =>
  y.a().flatMap(z => y.f(z) flatMap x.f).resume
```

I've got to get to work for now, but I'll follow up on this later.