Set up Machinist, use macros for Eq through Order syntax.
This commit adds cats-macros, upon which cats depends.
Right now, the only macros being used are Machinist's
operator macros, which are configured in cats.macros.Ops.

The punchline of doing this is that:

    def xyz[A: Eq](x: A, y: A): Boolean =
      x === y

is now exactly the same as:

    def xyz[A](x: A, y: A)(implicit A: Eq[A]): Boolean =
      A.eqv(x, y)

i.e. the syntax implicit conversions are completely erased.
This eliminates the performance penalty that syntax normally
implies. It also means the syntax classes don't need to be
value classes (or specialized) since they are not used at
runtime.