Added Reducible[_] type class.
This commit does a lot of related things. Here's a list:

 1. Removes existing foldRight, renames foldLazy to foldRight
 2. Makes Fold[_] instances a bit more powerful to work with
 3. Removes Comonad[OneAnd[?, F]] instance.
 4. Add some documentation.
 5. Some whitespace and formatting changes in a few files.

These changes are all possibly contentious.

(1) was intended to address the confusing bestiary of various
fold methods: foldLeft, foldRight, and foldLazy. Scala's eager
foldRight method is mostly-useless, which is why it seemed
reasonable to just remove it and replace it with our version.
I would be open to a different name -- I became convinced
foldLazy was a bad name but maybe I'm wrong?

(2) adds the .imap and .complete methods to Fold[_]. In
particular, .complete is something that almost anyone working
with Fold[_] will probably want.

(3) addresses a problem I noticed awhile ago: the Comonad
instance we had for OneAnd only dealt with the .head, dropping
the .tail on the floor. This is definitely wrong, although
we don't have a law to encode this yet. I noticed it while
working to have OneAnd implement Reducible.

(4) just adds a bit of documentation that I felt was missing.

(5) is arguably bad, but I tried to clean up some files that
I happened to be editing. I am torn between wanting to leave
lines alone (so that git blame is more useful and to preserve
the original author's intentions) and standardizing on a more
approachable style.