Document Lazy[_] and try to make rationale clearer.
This commit makes some changes that are aimed at trying to
clear up confusion around the Lazy[_] type. It's based on
my initial goals for the type -- we may decide we want it
to mean something else, in which case the docs will have to
change.

The only major code change was renaming `.force` to `.value`.
This is consistent with my interpretation (reifying lazy vals
and by-name params in the type system) and is also consistent
with scalaz.Name.