Introduce partialFold method to allow better composition.
The foldLazy method is written with a structure similar to
foldRight. However, it is written in a way that makes it
hard to compose multiple Foldable instances and still get
the correct behavior. This mean that CompositeFoldable
would not get the benefit of short-circuiting.

The partialFold method solves this problem. Since it returns
a Fold instance, calls can be chained while preserving the
correct semantics (i.e. short-circuiting as soon as Return
is seen).