Add Group instances for numbers.
In the long run Algebra will provide these instances,
but the situation right now is complex, so we will
paper things over here in Cats-land.

It also introduces a Show.ToStringShow[A] class that
can be used to provide Show[A] instances for concrete
types (this replaces many boilerplate definitions).