Lazy forms a valid monad
this also adds an Eq for lazy, however in order to test for equality it
has to force the lazy value, which feels "weird", but an Eq instance is
necessary to prove the laws. Not sure if perhaps the Eq instance should
be public or just used to validate the laws.