Make some changes to the Map[K, ?] instances.
This commit does a few different things:

 1. Changes the flatMap implementation
 2. Adds efficiency overrides for apply, apply2, map2
 3. Does some reformatting

The first is the most important. I don't think the
existing implementation is predictable -- the result
you get will be highly-dependent on the order that
the maps are iterated over. The new implementation
should behave the same no matter the order.

The second change is a trade-off. I feel like it is
worth it for us to make the standard instances a bit
more verbose to help improve users' performance.

The format changes were primarily around readability.
For example, I worried about people trying to read
two levels of string interpolation in Show. Other
parts were just personal taste (i.e. using a
non-anonymous class).