Add State/StateT
To avoid stack overflows, a couple design choices were made that make
State more complicated than it seems like it should be.

State wraps F[S => F[(S, A)]] instead of the more familiar S => F[(S, A)]. This allows for effective trampolining.

State[S, A] is a type alias for StateT[Trampoline, S, A] to avoid stack
overflows that are easy to hit if Id is used instead of Trampoline.

State is currently in its own module because it depends on Free (for
trampolining) but doesn't seem to belong in the Free module. Since it's
tough to avoid trampolining in FP Scala, we may want to reconsider
whether Free should be in core.

There is a brief ScalaDoc comment on StateT, but a proper markdown file
should probably be created. If this gets merged, I can send a follow-up
PR with some more documentation.