Add six new Comonad laws.
This commit adds some new laws, as well as adding an EqK[F[_]]
which makes it a bit easier to write tests. Like ArbitraryK,
EqK is defined in the laws module, rather than core.

It also adds some Eq and EqK instances that were previously
absent (the EqK instances are also in laws not core or std).