Define State companion in the package object
As it stands, `State` causes strange behaviors like State.apply working
the first time, but failing the second time to find apply.
@retronym pointed to me to SI-7139, which says that the REPL gets
confused about the type when there's a type alias and an object named
the same.

This adds the workaround to SI-7139 by moving State companion into the
package object.