Add a Bimonad law.
The law requires that extract(pure(a)) = a.

This commit also updates the Monad* tests to use the new EqK
typeclass, and also relocates some laws from ComonadLaws to
CoflatMapLaws.

There is a fair amount of plubming that had to change to get
all of this working. A really important but thankless task
will be to go through all of our tests and use EqK and
ArbitraryK where possible to create a more consistent
experience. This will only get harder once we have a new
ScalaCheck release and we have to worry about Cogen as well.