Add syntax tests.
This commit adds a file which tests that the syntax we
expect to provide is working as expected. It is not
complete, but currently tests the following type classes:

 - Eq
 - PartialOrder
 - Order
 - Semigroup
 - Group
 - Foldable
 - Reducible
 - Functor
 - Apply

While writing these tests I found two syntax traits that
were not being mixed into the "all" syntax, as well as
two operators with incorrect prototypes. Hopefully this
helps demonstrate why this is a good idea.