Add extract method to State instances
Similar to the extract method on the State companion object.

The name "extract" seems meaningful to me, but I realized that it
is a bit unfortunate that it's also the name of the method on
Comonad. Since State is implemented in terms of Trampoline, I thought
it might be kind of handy to have a runExtract method on StateT
that requires a Comonad and runs the State and extracts the value
in one go. This would make the name collision particularly unfortunate.

Anybody have any thoughts/ideas? One that comes to mind for me is
renaming the current State.extract to zoom.