Add forall and empty to Foldable
I also added tests/laws that forall and exists are lazy. This
laziness is probably more of a suggested implementation feature
rather than a _law_. However, since people will probably always
want their structures to follow this suggestion, I've added it to
the laws so mistakes are easily caught. I'm open to the idea that
this constraint shouldn't be enforced.