Major overhaul of Lazy[_], Foldable[_], etc.
This commit does a number of significant things:

 * Removes the Fold[_] implementation entirely.
 * Renames Lazy[_] to Eval[_]
 * Support monadic recursion over Eval[_] via trampoline
 * Remove partialFold &co
 * Update foldRight definition to be more standard
 * Fix documentation of Lazy[_], Foldable[_], etc.
 * Fix everything else to work with these changes

The trampoline used in Eval[_] is specialized and should be more efficient than
the standard encoding (although benchmarks are still needed to confirm this).
The semantics behind Eval[_] are hopefully a bit clearer, and the documentation
tries to make it clear exactly what one can expect.

I think this is a huge improvement over the status quo. The whole undertaking
was inspired by Miles' desire to have easier derivations for Foldable, so I
hope that this takes care of that. It also tries to address many of the
concerns around Lazy[_] that were brought up (many of which were ultimately
correct).

Review by @tpolecat, @milessabin, and anyone else who is interested.