Add Adelbert Chang to Cats maintainers.
Cody Allen suggested that since Adelbert has been very active
and helpful we should add him as a maintainer. I agree.

This PR adds him to the maintainers section in the README.
Assuming @adelbertc signs-off on this PR and no one objects
in the next few days, I'll add him as a committer.