Optimize Later[A].
This commit allows instances of Later[A] to "forget" the
thunk they use to lazily-evaluate their computation. This
means that expensive objects used to compute Later[A]
instances will be garbage-collected once the value is
computed.

Previously, a Later[A] instance would have held onto the
captured thunk, which in turn would have held onto its
captured objects. If those objects were large, this could
manifest as a larger-than-necessary memory footprint.