Update PR in response to comments.
This PR does a few things:

 * Renames Stream/StreamT to Streaming/StreamingT
 * Sets up some law-checking for Stream and StreamingT
 * Adds some type class instances
 * Fixes Stream#filter bug

One issue which was pointed out in the PR is that some
methods on StreamingT[F, ?] require F to be trampolined.
For types like Option which are not trampolined, this
will result in StackOverflowExceptions (and thus failed
law-checking) during some operations.

I don't have a 100% satisfactory answer to this right
now. I'd love to come up with a design that is completely
safe but don't see how it can be done in general.

Still left to do:

 * Update documentation
 * Comprehensive type class definitions for Streaming/StreamingT.
 * Do some benchmarking on Streaming/StreamingT overhead.
 * More unit tests for Streaming/StreamingT