Rename ADT nodes for Streaming and StreamingT.
 * This() becomes Cons()
 * Next() becomes Wait()
 * Empty() remains Empty()

There are a number of benefits to this change:

 1. `this` is a keyword, so could not be a method name
 2. This() doesn't necessarily communicate head+tail
 3. Next() doesn't necessarily mean anything.

When the ADT classes became private in StreamingT, the need
to create and name factory constructors prompted the creation
of StreamingT.cons and StreamingT.wait, which did a better
job of communicating what they did. Thus, those ended up
making good names for the ADT classes too.