Disable clean step in release to reduce memory pressure.
Currently I'm unable to get through a clean cross-build for
2.10 and 2.11 with both JVM and JS. Even using additional
memory (4G) I seem to run out late in the 2.11 build/test
after building and testing 2.10.

This suggests a memory leak in SBT or scalac. If we remove
the clean, I should be able to "stage" the release by
building ahead of time. The commit also removes parallel
execution from the JS step for the same reason.