Introduce cats-jvm and cats-js.
This commit does a bunch of things:

 * Adds a cats.macros.Platform object
 * Converts cats-macros to shared/jvm/js.
 * Converts other Cats modules to CrossType.Pure.
 * Moves JVM-specific code into cats-jvm.
 * Supports limited platform checking in laws/tests.

The new Platform object provides .isJvm and .isJs. These are
notable for being macros: they will be expanded at compile-time
to be either true or false. This allows scalac to eliminate
the "other" branch at compile-time, which will be useful to
us in laws/tests.

Moving forward, the goal is to only admit platform-specific code
(e.g. using java.io.*, or thread-blocking) into cats-jvm and
cats-js (which will not be required to provide the same APIs).
Other modules should be of CrossType.Pure (that is, they should
not have JVM- or JS-specific code in them).

The platform checking using .isJvm and .isJs should only be
used in cats-laws and cats-tests (although of course users of
the Cats library may use them however they like).