Adjust implicit priorities in OneAnd.
scala> implicitly[Functor[NonEmptyList]]
<console>:27: error: ambiguous implicit values:
 both method oneAndFunctor in trait OneAndInstances ...
 and method oneAndMonad in trait OneAndInstances ...
 match expected type cats.Functor[cats.data.NonEmptyList]

There were two bugs: one is that the Monad implicit
conflicts with the Functor implicit. The other is that
this construction doesn't work:

  object OneAnd extends OneAndInstances with OneAndLowPriority

If it DID work, this ordering would give priority to the
implicits found in OneAndLowPriority, so it's reversed - but
more to the point, implicit priority isn't based on linearization.
One of the traits has to be a subtrait of the other to have
higher priority, and as written they had no relationship.