Add an examples subproject.
The goal here is for people to be able to add
interesting code snippets here.

It's possible that many pieces of example code
which might live here could (or should) be expanded
into tutorials with narrative text. I still think
it might be smart to have a file like this, but
it wouldn't be hard to work the single example
I provided (dealing with Streaming) into a Tut
document.