Change MonadError from F[_, _] to F[_]
I was working with MonadError today and the `F[_, _]` approach made me use
quite a few type lambdas when I was working with types like Future/Task.
Since the error type is already encoded as another type parameter, as
far as I know we don't really lose anything by using a type constructor
with a single parameter for `F`.

It's quite possible that there's a reason for using `F[_, _]`. If anyone
knows of one, please speak up.