Rename *Aux helper classes to *Curried
We currently have a few helper classes that have an `Aux` suffix
for emulating currying type parameters. These helper classes have
an `apply` method which allows some type parameters of a function
to be inferred while others need to be specified. @tpolecat has
nice explanation of this trick here:

http://tpolecat.github.io/2015/07/30/infer.html

Unfortunately the suffix `Aux` is an overloaded term that is most
prominently used to provide terser syntax for type refinements, e.g.:

type Aux[F[_], A, B] = Coyoneda[F, A] { type Pivot = B }

I'm wondering if we should use `Curried` instead of `Aux` for classes
that enable type parameter currying. This could make the purpose of
those classes clearer to those who are unfamiliar with this trick.