XorT.toOption returns an OptionT
This makes `toOption` on `XorT[F, A, B]` return `OptionT[F, B]` instead
of `F[Option[B]]`. This makes it symmetrical with `toRight` and `toLeft`
on `OptionT`. My assumption here is that if you are working with `XorT`
instead of `F[Xor...]` that you would also prefer to work with `OptionT`
rather than `F[Option[...]]`. And if not, it's easy enough to call
`.value`.