Make a bunch of instances traits/classes private/sealed
This is related to #496, but it's a broader issue than just this change.

These instances traits/classes are used internally for code organization
and implicit prioritization, but they aren't really meant to be part of
the public API. Making them private/sealed may be helpful for binary
compatibility in the future. It also may be a helpful indicator to
newcomers which traits/classes are actually meant for consumption, so
they can ignore some of the noise.

So far, I have left the traits for orphans from `std` (such as
`StringInstances`), and I have left the omni-instance traits exposed.