Revert Free.foldMap regression
This reverts most of the changes from #702, because it appears to have
caused a regression in the case of Gosub. This commit fixes the
"mapSuspension id" unit test and the issue reported in #712.

The stack safety test is now failing. I've commented it out for now,
because an incorrectness bug seems to be worse than a stack safety
issue.