Remove Streaming
This is based on #858 and a discussion at dinner after the first day of the typelevel summit.

Per the discussion, Streaming will be moved to the dogs project.
StreamingT will be removed for now, and we can consider adding it to
dogs in the future if people find themselves wanting it.