Improve testing, remove lawless instance.
This commit fixes some problems which better testing
exposed. The biggest was that when we moved from the
"sparse vector" interpretation of maps to the "strict"
interpretation some of our instances became unlawful.
We can't define a group directly on maps because we
can't enforce the equivalence of "absent key" and
"present key with empty value".