Add Apply.map2Eval and allow traverse laziness
Fixes #513.

I think this should solve the majority of use-cases for lazy traversal.
One thing I noticed is that both `reduceRightToOption` and `traverse1_`
(which uses the former) seem to be impossible to implement lazily such
that they allow for short-circuiting. At least I couldn't figure out a
way. I left a note in the `traverse1_` scaladoc, but in practice I don't
anticipate this being a problem, because at the cost of just having an
`Applicative` instance as opposed to `Apply`, you can use `traverse_`
which is sufficiently lazy.