Replace 0.5.0 pre-release with more general notice
Cats `0.5.0` is up in Maven and there's already a `0.6.0-M1` release
up. With that in mind this message is a little unclear, and may be
suggesting to users that they should use previous releases
(e.g. `0.4.x`). Replacing the `0.5.0` warning with a more general
message about the state of Cats' development is clearer.