Provide concise .toString for free structures.
Previously calling .toString on a Free value would traverse
the entire ADT of case classes. In some cases this would
actually cause a stack overflow in the printing code. It
also represented a (small) law violation (you could use the
output to distinguish x and x.map(identity) in some cases).

This change replaces the output with opaque "Free(...)"
and "FreeApplicative(...)" strings.