Add FAQ section to site
This is just a start, and there are some more items that we will
probably want to add.

The import section probably is really a start to #912, and may end up
simply referencing a separate import document, as discussed in #936. But
I thought it would be nice to get something out there as an interim
resolution.