Reduce size of arbitrary ListWrappers
This is an experiment in response to #957. If this doesn't have much of
an effect on build times, then I think we can go ahead and close out
that ticket, since we are no longer seeing build timeouts. I suspect it
won't make much of a difference, because we already have changed to a
more performant JS environment and have reduced the scalacheck size
configuration in CatSuite.