Remove MonadFilter.filterM
Resolves #1054.

This simply removes `MonadFilter.filterM` to prevent confusion due to
this `filterM` being different than that from Haskell or Scalaz.

This version of `filterM` doesn't seem particularly useful to me. The
only type that comes to mind for which this might be useful is `Option`,
and it's pretty easy to come up with a more straightforward
implementation for the `Option` case. A more general solution for
`filterM` is provided by #1148, but so far there doesn't seem to be any
interest in that proposal, so I don't want it to hold up resolving