Don't run stack-safety tests for JS
This is a shot in the dark attempting to fix #1242.

The idea is that we can mark certain tests as slow and not run them when
running JS tests. Currently I've done this for all of the stack-safety
tests that I can find.

I don't really like this solution, but currently almost every PR is
resulting in a failed build, which is really problematic. One thing that
we could do is get rid of the `Platform.isJs` branch of the `slowTest`
implementation locally just before a release to make sure that these
tests all pass on JS. I think we only really run into these issues on
Travis CI.