Some attribution work
This is inspired by #1275. It doesn't include a CLA, and there is
most likely more work that should be done with respect to attribution,
but I'm hoping that this is a step in the right direction.

I've added a link to the scalaz contributors page within AUTHORS.md.
I'm not sure who all out of that should and would want to be directly listed
in the authors page for Cats. I've explicitly added Brian McKenna to the
authors page, as my understanding of [this
tweet](https://twitter.com/puffnfresh/status/762901748772057088) is that
he would like to be on there.