Drastically speed up Monoid[Map[K, V]] instance.
This commit fixes a number of problems I introduced with cats-kernel.

 1. Map's monoid was not overriding combineAll
 2. It was also doing the wrong thing for combine
 3. Monoid should have overridden combineAllOption

The performance for map's `combine` is 150x faster after this commit,
and `combineAll` is 780x faster. Rather than focusing on how good
these speed ups are, I want to apologize for how bad the
implementations were previously. The new `combine` is competitive with
other implementations, and `combineAll` is about x5 faster than the
equivalent code using combine.

(All figures using before/after numbers for the given benchmark.)

Thanks to Adelbert and Travis for noticing this problem. Also thanks
to Travis for producing this benchmark! Finally, thanks to Oscar for
helping me look at and fix this issue.