Roll-back #1301 and #1289 to allow 0.7.0 release.
Due to a communications error we merged some PRs which we had been
planning to keep out of master until after 0.7.0. Since one of them
(#1289) was large and long-running, it was difficult to undo this
merge in an automated way.

This commit represents the following:

 1. Downloaded patches for PRs 1301 and 1289 from Github
 2. Tried to reverse-apply them (-R)
 3. Fixed rejections/failures manually
 4. Got code compiling and tests passing

The code is now different than it was before the PRs were merged,
but things *seem* to be in a reasonable state. I'm keeping all these
changes in one commit so that it will be easier to "reverse" this
once 0.7.0 is relaesed.