Use scala versions define in .travis.yml
Using sbt-travisci 'crossScalaVersions' and 'scalaVersion' are defined
in terms of the scala versions defined in .travis.yml. Specifically, for
'scalaVersion' the last scala version is used (therefore either re-order
them or override it in build.sbt if it's not the default you want).

Also, define 'organization in ThisBuild' once instead of defining it in
'buildSettings' and appending it to every single project (in project
scope).

Finally, remove use of '%%' inside 'addSbtPlugin', it's overridden
anyways by 'addSbtPlugin'.