Resurrect deleted section about `Either`
Deleted section is still relevant - it was meant to present use of `Either` to reprsent failing computations. Variance
or not, it is more aout how to use this data type.
I tried to rewrote parts that are no longer true in presence of variance. Some other parts have been rearranged.