Partially revert 4530d3aa93131ec28096968a3c903ed1016dbf1b. Add back v… (#1506)
* Partially revert 4530d3aa93131ec28096968a3c903ed1016dbf1b. Add back variance on methods in EitherT and Either syntax.

* Add missing tests for Either syntax

* Fix the docs to match added variance to Either's enriched `flatMap`. `flatMap` now behaves identically to 2.12

* Resurrect deleted section about `Either`

Deleted section is still relevant - it was meant to present use of `Either` to reprsent failing computations. Variance
or not, it is more aout how to use this data type.
I tried to rewrote parts that are no longer true in presence of variance. Some other parts have been rearranged.