Use 2 workers in JVM build (#1673)
As far as I can tell, our CI build has 2 cores available on Travis. So I
suspect that 2 workers might be a sweet spot for our JVM testing. As I
understand it, the JS runtime is single-threaded, so we probably won't
benefit from additional workers on it (as our tests should be
CPU-bound).

I could be totally wrong about these assumptions, but I'm interested in
how the build times for this PR compare.