Improve StateT Arbitrary instance
As pointed out [here](https://github.com/typelevel/cats/issues/1640#issuecomment-298125906)

This reveals law violations in `StateT`. It looks like `flatMap`/`ap`
consistency does not hold, as well as `MonadCombine`
right-distributivity. It's not immediately clear to me how to fix this,
so I'm opening this up in case somebody else gets to it first.