Adding the origami project to the list of typelevel projects
This is a proposal to add `origami` to the list of Typelevel projects. However if I remember correctly one of the criteria for being part of the system is to have a compiler-checked documentation which doesn't exist anymore (it was there on a previous version of the library). I will add it and maybe then we can merge in this PR? Is anything else needed?