build: use curly braces for disabling tasks (#1876)
Using parentheses for disabling tasks with sbt 1.0 will issue the
following deprecation warning:
```
warning: Adaptation of argument list by inserting () is deprecated: this
is unlikely to be what you want.
        signature: DefinableTask.:=(v: S): sbt.Def.Setting[sbt.Task[S]]
  given arguments: <none>
 after adaptation: DefinableTask.:=((): Unit)
  publish := (),
          ^
```

This can be fixed with curly braces which is also perfectly fine with
sbt 0.13. So this can be merged right now and will make the actual
upgrade to sbt 1.0 easier.