Add val for "compile-time" sbt Configuration
This is another change that makes upgrading to sbt 1.0 easier
and which is compatible with sbt 0.13.

With sbt 1.0 the following line in `commonSettings`:
```
ivyConfigurations += config("compile-time").hide
```
leads to this exception:
```
[info] Loading settings from build.sbt,version.sbt ...
[info] Resolving key references (15580 settings) ...
[error] java.lang.IllegalArgumentException: requirement failed: id must be capitalized: commonSettings
[error]         at scala.Predef$.require(Predef.scala:277)
[error]         at sbt.librarymanagement.Configuration.<init>(Configuration.scala:19)
[error]         at sbt.librarymanagement.Configuration$.of(Configuration.scala:72)
[error]         at $61255add5b0123fd7403$.$anonfun$commonSettings$9(build.sbt:47)
```

The reason why `commonSettings` is used as `id` for this "compile-time"
configuration can be found in [this sbt-librarymanagement macro](https://github.com/sbt/librarymanagement/blob/53c80f076a5e9ab469037f352a4d68173140e216/core/src/main/scala/sbt/librarymanagement/ConfigurationExtra.scala#L109-L114).

As the macro suggests, extracting the defintion of the `Configuration`
and assigning it to a `val` fixes the error.

The "compile-time" configuration was added in
https://github.com/typelevel/cats/pull/1638 to remove simulacrum from
the dependencies in the POM. I therefore checked that simulacrum is
still not present in the POM after this change.